+++ CHANGELOG.rst
@@ -22,6 +22,7 @@ Features
   `Dougie Squire`_
 - Added ``dim`` and ``weights`` for probabilistic metrics. (:pr:`121`) `Aaron Spring`_
 - Added ``keep_attrs`` for all metrics. (:pr:`122`) `Andrew Huang`_
+- Added ranked probability score ``rps``. (:pr:`163`) `Aaron Spring`_
 
 Breaking Changes
 ----------------